app/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.7.0
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.4
config==0.5.1
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
distro==1.9.0
dnspython==2.5.0
dominate==2.9.1
email-validator==2.1.0.post1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.0
Flask-Github==0.1.3
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-mongoengine @ git+https://github.com/MongoEngine/flask-mongoengine@d4526139cb1e2e94111ab7de96bb629d574c1690
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
fonttools==4.47.2
GitHub-Flask==3.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.5
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.7.3
mongoengine==0.27.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.4
oauth2==1.9.0.post1
openai==1.13.3
packaging==23.2
pandas==2.0.3
pillow==10.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
rake-nltk==1.0.6
redis==5.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
visitor==0.1.3
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wordcloud==1.9.3
WTForms==3.1.2
zipp==3.17.0